Add build products to .gitignore.
Make a bunch of temporary changes to my BSD makefiles that you would not be
interested in reading about.

Add sources for a POSIX Threads-using C program, lpcq, that passes some numbers
from a producer thread to a consumer thread through a linked queue.  The
program accepts an option, -l, that enables synchronization of queue accesses
with a mutex.  RV-Predict detects races if lpcq is run without -l, and lpcq may
even corrupt the queue and timeout or crash.  Run with -l, the program always
finishes successfully, and RV-Predict detects no races.