Fix synchronization between rvp_wake_transmitter() and serialize(). Previously, serialize() prematurely stopped emptying ring buffers, so some rings would block forever.
Call all of the once-only initialization through rvp_init() instead of
having three separate once-only invocations.

Fix a couple of bugs involving thread 1's 'begin' op: 1) it was not
serialized at all, and after fixing that, 2) it was serialized after
other ops.

Take care not to record redundant 'switch' ops.  XXX It is necessary to
avoid more of these: you cannot have a 'switch T' before a 'begin T'.

Write trace files to ./rvpredict.trace (for now) instead of to /dev/null.
This is something I will control with an environment variable in the
near future.

Instead of opening the trace file in the serialization thread, try to
open the trace file before forking the thread, and exit early if the
file cannot be opened.  To help ensure that thread 1's 'begin' op is
written first to the trace file, flush thread 1's ring buffer before
starting the serialization thread.

Start a program, `rvpdump`, that dumps a trace file in human-readable
form.

Add to the next-generation runtime (ngrt) a type `delopts_t` to share
with rvpdump.

Refactor rvp_thread_flush_to_fd a bit to share writev() calls between
branches.

Fix a bug in rvp_ring_put_addr() where it wasn't iterating over the
uint32_t's comprising a uintptr_t, it was just writing the 0th one over
and over!

In rvp_ring_put_begin(), don't emit the new program counter between the
deltop and the thread ID, but emit it after the ID.

In the rvp_op_t definition, line up all of the enum values in the
same column.