Add a flag '-a' that makes rvsyms print the program counter / data address that had looked for.  This is useful for joining symbols with other data in scripts.
Change the walk-context member `residue` from from an int to a uint64_t,
since it should never go negative, and an int just isn't wide enough to
express some of the differences we see as a matter of course, especially
when there is not a match.

Make dwarf_walk_first()/dwarf_walk_next() track the depth of the
walk-context string stack and restore the old depth when making steps
upward and rightward in the debug-information tree.

Extract the subroutine dwarf_follow_attr_to_die() from
dwarf_follow_type_to_die(), and use it in dwarf_follow_spec_to_die(),
which looks up a DIE's specification DIE.  Use a specification DIE to
fill information about a C++ method, such as its name.

Produce line- and column-number information when it is available.
In check_line(), imitate what addr2line (part of elftoolchain) does to
produce the right line number when there is not an exact match of the
program counter we're searching for and the program counter of a line.

Use the filename from the line-number information instead of the name
of the current compilation unit to symbolize program counters.

Use a different output format for symbolizations.  I'm trying to make
it easier for a shell script to parse with grep and sed.

On standard input, parse {0x<address>}, [0x<address>], and []-variants
that tell one or two program counter/canonical frame address pairs,
and symbolize the given addresses.