Make git ignore a bunch of build products.
Add a rudimentary installer for RV-Predict/C, c-install/.

Add "interrupt service routine (ISR)" to the jargon file.

Stop using RV_ROOT in any of the scripts, since it complicates
installation for the user.  Update the manual to match.

Stub out some parts of the manual that need to be written.
Mention installation dependencies in the manual.

Don't build metadatadump and tracedump, they're not used any more.

Simplify: don't install checkJava and rv-predict scripts.  Instead,
put their functionality directly into rvpx.sh.

Unless RELEASE=yes is in the environment, build a stripped-down .jar
file with an embedded classpath that is suitable for a developer to run
in their local environment.

In ouat/, don't build the little-used stackstruct and walker.