Make git ignore *.deb.
Don't try to uninstall elftoolchain, it installs nothing.

Move the .jar file from lib/ to share/rv-predict-c/ like the Debian
package linter, lintian(1), prefers.

Update rvpa so that it runs the .jar from the new location.  Also,
instead of redirecting the prediction backend's stderr to stdout, swap
stderr and stdout.  That should work better with the current state of
Virgil's console progress indicator (which is still on another branch).

In maven/Makefile, avoid running mvn for the install target.  In this
way, if we're running 'mkcmake install' as root, which is necessary
for building a Debian binary package, we're not running all of that
Maven stuff.  Also, 'mkcmake install' is faster this way.