Print every progress indication with the same width so that one line does not intermingle with another.
Follow the last progress indication with a newline, so that the first
line of the first race report (or the "No races found." indication)
does not intermingle with the text of the last indication.

Insert a newline between the last progress indication and the race
reports.  (It would also look nice if we just erased the last progress
indication, instead.)

My code changes are kind of hackish, sorry.  Please feel free to handle
this another way.