Don't raise any signal on top of its handler.
Count the number of times that each signal handler finishes.

Add a routine __rvpredict_isr_fire(isr) that I can use in ServoModule
to replace direct calls to an ISR by the application thread.
__rvpredict_isr_fire(isr) raises the signal whose handler is `isr` and
waits with our interrupt-simulation signals unblocked until `isr` has run.
In this way, it looks in our trace like `isr` was entered asynchronously.
Otherwise, it looks like the thread entered `isr` synchronously, so we
won't predict a data race.

Replace a bunch of instances of SIGRTMIN + i with a temporary variable,
signum, so that it's easier to read and check the code.