Add an `rvp-` prefix to the interrupt service routine (ISR) function annotation.
Add a new annotation, __rv_register, for use with variables that
simulate device registers.  When our instrumentation sees a a store
to an N-byte-wide device register, it replaces the store with a
call to __rvpredict_pokeN(), which performs an uninstrumented write.
A load from an N-byte-wide device register is replaced with a call to
__rvpredict_peekN(), which performs an uninstrumented read.  In this
way, when a test fixture initializes simulated device registers before
it runs some application code that subsequently reads those registers,
no happens-before relation is established between the test-fixture stores
and the application loads.

Add nascent code (currently disabled) for initiating data-race prediction
directly from the instrumented binary instead of through rvpx/rvpa.