Overhaul my scheme for walking signal rings---and get most of the bugs out, I hope!
Fix a couple of bugs that would lead to deadlock: intern_sigset() can
be called from signal context, so it must not acquire a lock.  Now it
does not.  Also, to avoid deadlock, holders of the thread lock need
to block signals: a signal handler that interrupts a thread that holds
the thread lock may ask for ring service from the serialization thread,
but the serialization thread must hold the thread lock while it performs
that service -> deadlock.