Interpret format specifiers %n %u %t %p in the environment variable RVP_TRACE_FILE, interpolating the executable filename, the username, approximate trace-file creation time (seconds since the UNIX epoch began), or the PID, respectively.  Describe the format specifiers in the manual page for `rvpc`.
Make rvpa accept a second argument, the name of the trace file
to read.  In this way we can analyze a trace file that's not called
`./rvpredict.trace`.

Add some tests for `rvpa` that make sure that it detects & correctly
reports too many arguments and too few, and that it uses the right
argument as a trace filename whether it is supplied one or two arguments.

Make the Java backend treat the argument of --llvm-predict as a trace
filename instead of a directory.  For now, keep writing `debug.out` and
`result.txt` in "beside" the trace file---i.e., in the same directory.

Looks like I have accidentally committed a couple of scripts, `stdout`
and `swap.sh`, that I do not mean to.  I will fix that in the next commit.