Stop the interrupt simulation from tracing inconsistent signal masks. The interrupt simulation was always blocking a number of real-time signals without recording it in the trace.  Fortunately, the resulting inconsistency in the trace was revealed when assertions failed in the analysis backend.
Now, the interrupt simulation traces changes to the signal mask.
However, it avoids cluttering the trace with signal-mask events unless
it is actually operating: it only blocks a real-time signal if it's
establishing a simulated interrupt on that signal.

While I am here, in rvp_change_sigmask(), reactivate some diagnostic
code that raises SIGTRAP if there is a discrepancy between the signal
mask that is expected (given the prior signal mask, the operation and
operands) and the actual signal mask.