Pipe/named-FIFO compatibility and statistics.
Make the runtime work more effectively with named FIFOs and pipes, and
add an environment variable, RVP_TRACE_FIFO, that tells a named FIFO that
should receive a trace.

Start to count some statistics, and make the runtime print them when it
receives SIGINFO (on systems where SIGINFO is defined) or SIGPWR (all other
systems).

Drop a lock before calling errx(3) to prevent a hang.