Add an example, `examples/interrupts/sim/`, that uses the interrupt simulation.
Add to the `signal` example an option, `-r #`, that tells how many
signals may stack up on a signal.  `-r 0` is the default: a signal may
only interrupt a thread.  Since signals interrupting signals produce
a data-race report that we cannot symbolize, I haven't documented this
option, yet.

For the runtime, use the C options `-pedantic -Wmissing-prototypes`.
Add some #includes and sprinkle `static` so that compilation
succeeds.

Inspired by a client visit, improve the interrupt simulation: add
__rvpredict_splhigh() that blocks all interrupts and returns the current
interrupt priority, and __rvpredict_splx() that restores a prior interrupt
priority.  These work analogously to splhigh() an splx() in *BSD.

Respect the priority of static interrupts established using
__rvpredict_interrupt_handler() so that when an interrupt lowers the
interrupt priority with __rvpredict_splx(), the interrupt simulation only
triggers interrupts with higher priority than the one that currently runs.
In this way we avoid an infinite loop/recursion of interrupts.

Fix a bug where the `t_intrmask` member of `rvp_thread_t` did not
track the change to a signal mask that happens in a signal handler.
On entry to our signal-handler wrapper, set in `t_intrmask` the mask
bits established by sigaction(2); restore `t_intrmask` on the way out
of the signal-handler wrapper.

XXX Right now it's possible for signals to "stack up" before we ever enter
the wrapper, so the mask in `t_intrmask` may have fewer bits set than
the actual mask.  I.e., the actual mask is `t_intrmask | x` for some `x`
(possibly `x == 0`).  It's not clear to me whether the analysis backend
will tolerate the discrepancy.

Write a note to myself about sigaction(2) needing to be async-signal-safe.