In block_signals(), initialize the sigset_t before adding signals to it.
Instead of blocking signals in the application process, ignore them,
because the ignores won't be reset by fork+exec.  Now the `cancellation`
regression test should pass, and hitting Control-C to send `SIGINT` to
a process should cancel the application without stopping the analysis
from running.

Check the status of the analysis process and squeak on `stderr` if
it exited on a signal.