Fix race condition: hold the ring lock while (re-)checking the ring-full condition, waking the serializer, and waiting for the ring-full condition to subside.  Refactor rvp_ring_await_nempty() in support of this.
Assert here and there that locks are held.

Do not hold `thread_mutex` while flushing the rings.  In this
way, we avoid a lock-order reversal between the ring mutexes and
`thread_mutex`.  Also, a thread that needs to wake the serializer can
acquire `thread_mutex` and increase `nwake` while the serializer is
still busy; previously, a thread had to wait until the serializer was
idle to increase `nwake`, which is not efficient.