Increase the global generation number, if it still equals the local generation number, every 512 words written to the local event ring. Do not increase the global generation number in the serialization thread any longer, now that the individual threads take care of it.  Increasing the global generation every 512 words made the overhead unmeasurable.
Embed the ring content, `r_items`, into the `rvp_ring_t`, if EMBED_RING
is #defined.  Make EMBED_RING the default.

Wrap `rvp_ring_t.r_last` in accessor functions, `rvp_ring_last()` and
`rvp_ring_last_const()`, and use those instead of accessing `r_last`
directly.  If NO_RING_LAST is #defined, then don't use the pointer
`rvp_ring_t.r_last` to the last element on the ring---i.e., the one
before wrapping back to item 0.  Instead, just compute the last item
when it is needed.  I don't enable NO_RING_LAST by default because it
seems that loading the pointer is faster than computing it, *shrug*.

In `trace_store8`, make the code that avoids overrunning the end
of the ring buffer simpler and faster.  Also, don't detect overrun
RVP_BUF_NITEMS - 1 ahead, but wait until only `slots_per_store - 1`
items ahead; `slots_per_store` is the number of slots it takes to hold
the 8-byte wide store in the common case.

Just load `r->r_producer` once near the top of `trace_store8` for a
measurable performance increase.

In the `trace_store8` slow path, also increase the global generation
number when it is time.

Increase the number of bytes on the rings to 384 * 4096 B = 1536 kB and
ask the serializer for service when the ring is 3/4 full instead of 7/8.
This seems to be the smallest ring size and service threshold necessary
for top performance on `rvwork-1`.

Arrange for members `r_tid` and `r_idepth` of `rvp_ring_t` to be in a
single, read-mostly cacheline that is not falsely shared with ring items
or other frequently-written content.  This helped a bit with performance.

Add a branch hint to `rvp_ring_request_service`.  It actually seems to
help performance a bit.