Read the ring pointer, `rvp_thread_local.tl_intr_ring`, in `__rvpredict_store8` before calling `trace_store8` to save us about 0.6 cycles.  My hunch is that the read completes concurrently with some non-dependent instructions, so the sooner we issue it, the less time `trace_store8` must stall before using it.
It may pay to load the ring pointer just once in each instrumented
function, and then to pass it to users like `__rvpredict_store8`.
We'd probably get essentially the same benefit using LTO, if we could get
LTO to work.  We could run into problems if the ring changes because, say,
`fork(2)` is called.